% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclid.R
\name{Euclid}
\alias{Euclid}
\title{Calculation of the Euclidean Distance}
\usage{
Euclid(x, y, w, squared = FALSE)
}
\arguments{
\item{x}{A set of coordinates in the form of a \code{\link{matrix}} or
\code{\link{data.frame}}.}

\item{y}{An optional second set of coordinates in the same forms as \code{x}.}

\item{w}{An optional set of weights to be applied on the coordinates.}

\item{squared}{Should the squared Euclidean distances be returned (default:
FALSE).}
}
\value{
A \code{\link{dist}-class} object or, whenever \code{y} is provided,
a \code{\link{matrix}} with as many rows as their are rows in \code{x} and as
many columns as there are rows in \code{y}.
}
\description{
Function \code{Euclid} carries out the calculation of pairwise Euclidean
distances within a set of coordinates or between two sets thereof, with
optional weights.
}
\details{
When only one set of coordinates is given to the function (ie. when
\code{y} is omitted), the function returns the pairwise distances in the form
of a \code{\link{dist}-class} object representing a lower-triangle matrix. If
weights are omitted, the result is identical to that produced by
\code{\link{dist}} with argument \code{method = "euclidean"} (the function's
default). When weights are provided, the weighted Euclidean distance is
calculated.

The standard \code{R} function used to calculate the Euclidean distance
(\code{\link{dist}}), only allows one to calculate pairwise distances between
the rows of a single matrix of Cartesian coordinates and return a
\code{\link{dist}}-class object, which is a one-dimensional array meant to be
interpreted as a lower-triangular matrix. In addition to allowing one to
calculate weighted Euclidean distances, function \code{Euclid} can also be
provided two data matrices (\code{x} and \code{y}) and output a rectangular
matrix of the Euclidean distances between their rows.
}
\examples{
##
### First example
##
## [Examples here...]

}
\references{
[To be included...]
}
\seealso{
The \code{\link{dist}-class} and associated methods.
}
\author{
Guillaume Guénard \email{guillaume.guenard@umontreal.ca}
}
