% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting-functions.R
\name{weighting-functions}
\alias{weighting-functions}
\alias{wf.sqrd}
\alias{wf.RBF}
\alias{wf.PCNM}
\alias{wf.binary}
\alias{wf.Drayf1}
\alias{wf.Drayf2}
\alias{wf.Drayf3}
\title{Weighting Functions for Spatial Eigenvector Map}
\usage{
wf.sqrd(d)

wf.RBF(d, wpar = 1)

wf.PCNM(d, boundaries, wpar = 4)

wf.binary(d, boundaries)

wf.Drayf1(d, boundaries)

wf.Drayf2(d, boundaries, wpar = 1)

wf.Drayf3(d, boundaries, wpar = 1)
}
\arguments{
\item{d}{A triangular (\code{\link{dist}}-class) or rectangular geographic
distance matrix produced by \code{\link{dist}}, \code{\link{Euclid}}, or
\code{\link{geodesics}}.}

\item{wpar}{Where applicable, a parameter controlling the shape of the
spatial weighting function.}

\item{boundaries}{Where applicable, a two-element numeric vector containing
the lower and upper threshold values used to obtain the connectivity matrix.
(see details).}
}
\value{
A \code{\link{dist}-class} object when argument \code{d} is a
\code{\link{dist}-class} object or a rectangular matrix when argument
\code{d} is a rectangular matrix, either one with the weights as its values.
}
\description{
A set of common distance weighting functions to calculate spatial eignevector
maps using function \code{\link{eigenmap}}.
}
\details{
These functions are meant primarily to be called within functions
\code{\link{eigenmap}} and \code{\link{eigenmap.score}}. In
\code{\link{eigenmap}}, argument \code{d} is a lower-triangular
\code{\link{dist}}-class object and the resulting lower-triangular weight
matrix is used in calculating the spatial eigenvector map. In
\code{\link{eigenmap.score}}, \code{d} is a rectangular matrix of the
distances between a set of arbitrary locations (rows) and reference locations
(columns; the locations for which the the spatial eigenvector map has been
built and the resulting rectangular weight matrix is used to calculate
spatial eigenfunction values. These values allow one to use the spatial
information of a data set for making predictions at arbitrary values.

\code{Wf.sqrd} (default value) consists in taking
\eqn{w_{i,j} = -0.5*d_{i,j}} and does not involve any trunction.

\code{Wf.RBF} consists in taking \eqn{w_{i,j} = exp(-wpar*d_{i,j}^2)} and
does not involve any trunction.

\code{Wf.binary} the spatial weighting matrix is simply the connectivity
matrix,

\code{Wf.PCNM} is \eqn{a_{i,j} = 1 - (d_{i,j} / (wpar*boundaries_2))^2}
 
\code{Wf.Drayf1} is \eqn{a_{i,j} = 1 - (d_{i,j} / d_{max})} where \eqn{d_max}
is the distance between the two most distant locations in the set,

\code{Wf.Drayf2} is \eqn{a_{i,j} = 1 - (d_{i,j} / d_{max})^{wpar}},

\code{Wf.Drayf3} is \eqn{a_{i,j} = 1 / d_{i,j}^{wpar}}.

Functions \code{Wf.Drayf1}, \code{Wf.Drayf2}, and \code{Wf.Drayf3} were
proposed by Dray et al. (2006) and function \code{PCNM} was proposed by
Legendre and Legendre (2012).

The \code{Wf.sqrd} weighting approach is equivalent to submitting the
elementwise square-root of the distance matrix to a principal coordinate
analysis. It was proposed by Diniz-Filho et al. (2013) and is equivalent, for
evenly spaced transect or surfaces (square or rectangle), to using the basis
functions of type II discrete cosine basis transforms. However, this fact has
not been reported by Diniz-Filho et al. (2013).

When calculating the connectivity matrix, pairs of location whose distance to
one another are between the boundary values (argument \code{bounraries}) are
considered as neighbours (\eqn{b_{i,j}=1}) whereas values located below the
minimum and above the maximum are considered as equivalent or distant,
respectively (\eqn{b_{i,j}=0} in both cases).

User may implement custom weighting functions. These functions must at the
very least have an argument \code{d}, and can be given arguments
\code{boundaries} and \code{wpar}. Argument \code{wpar} may be a vector with
any number of elements, but the function definition must be provided with
default values. 

User-provided weighting functions with a \code{wpar} argument must come with
a valid default value for that parameter since \code{\link{eigenmap}} may
internally call it without a formal value.
}
\section{Functions}{
\itemize{
\item \code{wf.sqrd}: Principal coordinates of the square-root distance matrix (Diniz-Filho et al.
2013).

\item \code{wf.RBF}: Radial basis functions with the observations as the kernels.

\item \code{wf.PCNM}: Borcard & Legendre's (2002) principal coordinates of the neighbour matrix
approach.

\item \code{wf.binary}: Dray et al. (2006) Moran's eigenvector maps (distance-based binary
connections without continuous weighting of the neighbours).

\item \code{wf.Drayf1}: Dray et al. (2006) Moran's eigenvector maps (distance-based binary
connections with continuous weighting of the neighbours: f1).

\item \code{wf.Drayf2}: Dray et al. (2006) Moran's eigenvector maps (distance-based binary
connections with continuous weighting of the neighbours: f2).

\item \code{wf.Drayf3}: Dray et al. (2006) Moran's eigenvector maps (distance-based binary
connections with continuous weighting of the neighbours: f3).
}}

\examples{
locations <- c(1,2,4,7,10,14,17,21)
D <- dist(locations)
wf.sqrd(D)
wf.RBF(D, wpar = 0.1)
wf.binary(D, c(0,5))
wf.PCNM(D, c(0,5))
wf.Drayf1(D, c(0,5))
wf.Drayf2(D, c(0,5), 0.5)
wf.Drayf3(D, c(0,5), 0.5)

emap <- eigenmap(D, locations, wf.Drayf2, c(0,5), 0.5)
emap

emap <- eigenmap(D, locations, wf.Drayf3, c(0,5), 0.25)
emap

emap <- eigenmap(D, locations, wf.RBF, wpar = 0.1)
emap

}
\references{
Borcard, D. and Legendre, P. 2002. All-scale spatial analysis of ecological
data by means of principal coordinates of neighbour matrices. Ecol. Model.
153: 51-68

Diniz-Filho, J. A. F.; Diniz, J. V. B. P. L.; Rangel, T. F.; Soares, T. F.;
de Campos Telles, M. P.; Garcia Collevatti, R. and Bini, L. M. 2013. A new
eigenfunction spatial analysis describing population genetic structure.
Genetica 141:479-489.

Dray, S.; Legendre, P. and Peres-Neto, P. 2006. Spatial modelling: a
comprehensive framework for principal coordinate analysis of neighbor
matrices (PCNM). Ecol. Modelling 196: 483-493

Legendre, P. and Legendre, L. 2012. Numerical Ecology, 3rd English edition.
Elsevier Science B.V., Amsterdam, The Netherlands.
}
\author{
\packageAuthor{codep}
Maintainer: \packageMaintainer{codep}
}
